Move TestAction class to 'actions' package.
